Add react router and clean up
