+++ app/adapter/web/package.json
@@ -13,6 +13,7 @@
     "graphql-tag": "^2.10.1",
     "material-design-icons": "^3.0.1",
     "qrcode": "^1.4.1",
+    "react-router-dom": "^5.0.1",
     "recaptcha-v3": "^1.5.0"
   },
   "scripts": {
@@ -47,6 +48,7 @@
     "@types/qrcode": "^1.3.3",
     "@types/react": "^16.8.24",
     "@types/react-dom": "^16.8.5",
+    "@types/react-router-dom": "^4.3.5",
     "node-sass": "^4.12.0",
     "prettier": "1.18.2",
     "react": "^16.8.6",
