+++ app/adapter/web/src/component/pages/shared/Footer.scss
@@ -1,7 +1,7 @@
-@import "styles/dimension";
-@import "styles/font";
-@import "styles/color";
-@import "styles/easing";
+@import "../../styles/dimension";
+@import "../../styles/font";
+@import "../../styles/color";
+@import "../../styles/easing";
 
 footer {
   @include default-font();
