+++ app/adapter/web/src/component/pages/shared/Footer.tsx
