+++ app/adapter/web/src/component/pages/shared/Header.scss
@@ -1,5 +1,5 @@
-@import "styles/font";
-@import "styles/dimension";
+@import "../../styles/font";
+@import "../../styles/dimension";
 
 header {
   @include default-font();
