+++ app/adapter/web/src/component/pages/shared/promos/ExtPromo.scss
@@ -1,0 +1,22 @@
+@import "../../../styles/font";
+
+.ext-promo {
+  font-weight: $font-weight-medium;
+  padding: 14px 20px;
+
+  @include default-font();
+
+  text-align: center;
+  color: black;
+
+  a {
+    font-weight: $font-weight-semi-bold;
+    padding: 3px 10px;
+    text-decoration: none;
+    color: black;
+
+    &:hover {
+      opacity: 0.9;
+    }
+  }
+}
