+++ app/adapter/web/src/component/styles/color.scss
@@ -1,1 +1,1 @@
-$color-primary : #E91E63;
+$color-primary : #e91e63;
