+++ app/adapter/web/src/component/styles/corner.scss
@@ -1,3 +1,3 @@
 @mixin round-corner {
-  border-radius : 4px;
+  border-radius: 4px;
 }
