+++ app/adapter/web/src/component/styles/icon.scss
@@ -1,26 +1,27 @@
 @import '~material-design-icons/iconfont/material-icons.css';
 
 .material-icons {
-  font-family             : 'Material Icons';
-  font-size               : 24px; /* Preferred icon size */
-  font-weight             : normal;
-  font-style              : normal;
-  line-height             : 1;
-  display                 : block;
-  white-space             : nowrap;
-  letter-spacing          : normal;
-  text-transform          : none;
-  word-wrap               : normal;
-  direction               : ltr;
+  font-family: 'Material Icons', sans-serif;
+  font-size: 24px; /* Preferred icon size */
+  font-weight: normal;
+  font-style: normal;
+  line-height: 1;
+  display: block;
+  white-space: nowrap;
+  letter-spacing: normal;
+  text-transform: none;
+  word-wrap: normal;
+  direction: ltr;
 
   /* Support for all WebKit browsers. */
-  -webkit-font-smoothing  : antialiased;
+  -webkit-font-smoothing: antialiased;
+
   /* Support for Safari and Chrome. */
-  text-rendering          : optimizeLegibility;
+  text-rendering: optimizeLegibility;
 
   /* Support for Firefox. */
-  -moz-osx-font-smoothing : grayscale;
+  -moz-osx-font-smoothing: grayscale;
 
   /* Support for IE. */
-  font-feature-settings   : 'liga';
+  font-feature-settings: 'liga';
 }
