+++ app/adapter/web/src/component/ui/Button.scss
@@ -1,6 +1,6 @@
-@import "styles/corner";
-@import "styles/color";
-@import "styles/font";
+@import "../styles/corner";
+@import "../styles/color";
+@import "../styles/font";
 
 button {
   font-size: 16px;
