+++ app/adapter/web/src/component/ui/Notice.scss
@@ -1,7 +1,7 @@
 @import "../styles/font";
 
 .notice {
-  transform        : translateZ(100px);
-  background-color : white;
-  box-shadow       : 0 1px 3px rgba(0, 0, 0, 0.12), 0 1px 2px rgba(0, 0, 0, 0.24);
+  transform: translateZ(100px);
+  background-color: white;
+  box-shadow: 0 1px 3px rgba(0, 0, 0, 0.12), 0 1px 2px rgba(0, 0, 0, 0.24);
 }
