+++ app/adapter/web/src/component/ui/Section.scss
@@ -1,5 +1,5 @@
-@import "styles/font";
-@import "styles/dimension";
+@import "../styles/font";
+@import "../styles/dimension";
 
 .section {
   @include default-font();
