+++ app/adapter/web/src/component/ui/Section.tsx
