+++ app/adapter/web/src/index.scss
@@ -1,5 +1,5 @@
 @import url('https://fonts.googleapis.com/css?family=Source+Code+Pro:200,300,400,500,600,700,900');
 
 body {
-  margin : 0;
+  margin: 0;
 }
