+++ app/adapter/web/src/index.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import './index.scss';
-import App from './component/App';
+import { App } from './component/App';
 import * as serviceWorker from './serviceWorker';
 
 import { CaptchaService } from './service/Captcha.service';
